simplified the app
