+++ frontend/src/lib/config.json.example
@@ -1,13 +1,1 @@
-{
-        "firebaseConfig":{
-                "apiKey": "<api-key>",
-                "authDomain": "<auth-domain>",
-                "projectId": "<project-id>",
-                "storageBucket": "<storage-bucket>",
-                "messagingSenderId": "<messaging-sender-id>",
-                "appId": "<app-id>",
-                "measurementId": "<measurement-id>"
-            },
-            "recaptchaSiteKey": "<recaptcha-site-key>"
-        
-}
+{}
