+++ shared/schemas/index.ts
@@ -2,7 +2,5 @@
 export * from './common';
 export * from './business';
 export * from './customer';
-export * from './inventory';
 export * from './order';
-export * from './cashflow';
 export * from './event';
